requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
python-dotenv
pydantic